+++ packages/static/src/types.ts
@@ -15,8 +15,7 @@ export interface FallbackDescriptor {
 export type GetRouterOptions = () => RouterOptions;
 
 export interface StaticOutput {
-  render: (emitted: Emitted) => any;
-  insert: (markup: any, emitted?: Emitted) => string;
+  render: (emitted: Emitted) => string;
   dir: string;
   redirects?: boolean;
 }